Moved documentation utility commands to a folder
